+++ src/app/[locale]/(adm)/dashboard/admin/page.tsx
@@ -1,6 +1,8 @@
+import { Edit, Mail, View } from "lucide-react";
+
 import { Link } from "~/core/link";
+import { sendEmails } from "~/core/mail/resend";
 import AdminProductsManagement from "~/core/trpc/tanstack/products-admin";
-import { Icons } from "~/islands/icons";
 import { GeneralShell } from "~/islands/wrappers/general-shell";
 import { redirect } from "~/navigation";
 import Creating from "~/plugins/builder/creating";
@@ -11,6 +13,14 @@ import ButtonSetPrivilegies from "./_islands/privilegies";
 export default async function AdminMainPage() {
   const session = await getServerAuthSession();
   if (!session) return redirect("/auth");
+
+  // TODO: Implement email system
+  /* const email = "blefnk@gmail.com";
+  const firstName = "Nazarii";
+  enum Templates {
+    Onboard = "ONBOARD",
+  }
+  await sendEmails([email], { firstName }, Templates.Onboard); */
 
   return (
     <GeneralShell>
@@ -42,21 +52,29 @@ export default async function AdminMainPage() {
         <section className="mb-12 space-y-4">
           <h2>Create New Post</h2>
           <Link href="/blog/new" variant="secondary">
-            <Icons.edit className="mr-2 h-4 w-4" aria-hidden="true" />
+            <Edit className="mr-2 h-4 w-4" aria-hidden="true" />
             Create post
           </Link>
           <br />
           <Link href="/blog" variant="secondary">
-            <Icons.view className="mr-2 h-4 w-4" aria-hidden="true" />
+            <View className="mr-2 h-4 w-4" aria-hidden="true" />
             See all posts
           </Link>
         </section>
 
         <section className="mb-12 space-y-4">
           <h2>Edit User Privilegies</h2>
-          {/* eslint-disable-next-line jsx-a11y/aria-role */}
-          <ButtonSetPrivilegies userId={session.id} role="user" />
+          <ButtonSetPrivilegies userId={session.id} newRole="user" />
         </section>
+
+        {/* TODO: Implement email system */}
+        {/* <section className="mb-12 space-y-4">
+          <h2>Test Email Delivery</h2>
+          <Link href="/api/mail" variant="secondary">
+            <Mail className="mr-2 h-4 w-4" aria-hidden="true" />
+            Send email
+          </Link>
+        </section> */}
       </section>
 
       <hr />