+++ src/data/meta/index.ts
@@ -1,2 +1,2 @@
 export * from "./default";
-export * from "./extend";
+// export * from "./extend.txt";